requirements.txt
fastapi
uvicorn[standard]
asyncpg
supabase
python-multipart
python-jose[cryptography]
passlib[bcrypt]
bcrypt==3.2.0
pydantic[email]
celery
redis
pandas
numpy
scikit-learn
pulp
requests
geopy
google-genai
python-dotenv
